% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/current_URL.R
\name{current_URL}
\alias{current_URL}
\title{Get the current URL of active the WebDriver client}
\usage{
current_URL(WebDriver = remDr)
}
\arguments{
\item{WebDriver}{Active binding object containing the active WebDriver.}
}
\description{
A very simple wrapper for client$getCurrentURL that returns the current URL
}
