% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dlmet_download.R
\name{dlmet_download}
\alias{dlmet_download}
\title{Download files from a list of file names.}
\usage{
dlmet_download(
  filenames,
  page_url,
  dl_directory,
  verbose = TRUE,
  suppress_download_popups = FALSE
)
}
\arguments{
\item{filenames}{a matrix or dataframe containing the filenames of the files to be downloaded.}

\item{page_url}{the URL of the page from which the files are to be downloaded}

\item{dl_directory}{the destination folder. Where do you want the files to go?}

\item{verbose}{TRUE/FALSE display prompt to continue with operation.}

\item{suppress_download_popups}{TRUE/FALSE to stop download.file popups}
}
\description{
Download files using a list returned by other dlmet functions.
}
