% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_met_filenames.R
\name{get_met_filenames}
\alias{get_met_filenames}
\title{Scrape the filenames of the designated met type from the NOAA ftp server.}
\usage{
get_met_filenames(
  met_type = "gdas1",
  target_year = NA,
  verbose = TRUE,
  chromepath = "C:/Program Files/Google/Chrome/Application/"
)
}
\arguments{
\item{met_type}{type of NOAA met data (character string; either "gdas1" or "reanalysis").}

\item{target_year}{year of data frame subset, if any.}

\item{verbose}{TRUE/FALSE for window prompt to continue with function.}
}
\value{
A data frame containing all the filenames in the specified NOAA ftp met directory.

A data frame containing filenames from the target_year (if !is.NA).
}
\description{
Use RSelenium and Google Chrome to initiate a remote WebDriver, navi
gate to the NOAA ftp server, and scrape the file names of the specified met type.
}
\examples{
get_met_filenames(met_type = "reanalysis", target_year = "2020")
}
