% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grab_objects_onpage.R
\name{grab_objects_onpage}
\alias{grab_objects_onpage}
\title{List objects on the current NOAA ftp archive page.}
\usage{
grab_objects_onpage(object_type = "all", trim = FALSE, return = TRUE)
}
\arguments{
\item{object_type}{Scrape names/details of type 'file','folder', or 'all'.}

\item{trim}{TRUE/FALSE to remove columns other than filenames.}

\item{return}{TRUE/FALSE to return the scraped data.}
}
\value{
A data frame containing all the filenames of object_type from the current client web page.
}
\description{
Interrogate/list contents of the table on the current NOAA ftp server page.
}
\examples{
list_objects_onpage(object_type = "file", output_names = FALSE)
}
