% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initiate_client.R
\name{initiate_client}
\alias{initiate_client}
\title{Initialise an RSelenium WebDriver client with Chrome}
\usage{
initiate_client(
  verbose = FALSE,
  chromepath = "C:/Program Files/Google/Chrome/Application/",
  terminate = FALSE,
  init_url = "ftp://arlftp.arlhq.noaa.gov/archives/"
)
}
\arguments{
\item{verbose}{TRUE/FALSE generate query window prior to running.}

\item{chromepath}{directory of google chrome install.}

\item{terminate}{TRUE/FALSE terminate client immediately after starting.}

\item{init_url}{URL that the client will navigate to upon launching.}
}
\description{
Start up a remote WebDriver with RSelenium. This will check to confirm
everything is working properly. Option terminate immediately.
}
