% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grab_objects_onpage.R
\name{grab_objects_onpage}
\alias{grab_objects_onpage}
\title{List objects on a NOAA ftp archive page.}
\usage{
grab_objects_onpage(object_type = "all", output_names = TRUE)
}
\arguments{
\item{object_type}{get object names/details of type 'file','folder', or 'all'.}

\item{output_names}{TRUE/FALSE store object details as data frame.}
}
\value{
A data frame containing all the filenames in the specified NOAA ftp met directory.

A data frame containing filenames from the target_year (if !is.NA).
}
\description{
Interrogate/list contents of table on a NOAA ftp server page.
}
\examples{
list_objects_onpage(object_type = "folder", output_names = FALSE)
}
