% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initiate_client.R
\name{initiate_client}
\alias{initiate_client}
\title{Initialise an RSelenium WebDriver client with Chrome}
\usage{
initiate_client(
  WebDriverName = "remDr",
  chromepath = "C:/Program Files/Google/Chrome/Application/",
  terminate = FALSE,
  verbose = FALSE,
  port = 4567L,
  init_url = "ftp://arlftp.arlhq.noaa.gov/archives/"
)
}
\arguments{
\item{WebDriverName}{Name of the webdriver to be created.}

\item{chromepath}{Directory of google chrome Application folder.}

\item{terminate}{TRUE/FALSE to terminate client immediately after starting.}

\item{verbose}{TRUE/FALSE to generate query window prior to running.}

\item{port}{Which port to use for the remote driver? Shares RSelenium's default of 4567L.}

\item{init_url}{URL that the client will navigate to upon launching. NOAA ARL archive by default.}
}
\description{
Start up a remote WebDriver with RSelenium. Option to terminate immediately.
}
