% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terminate_client.R
\name{terminate_client}
\alias{terminate_client}
\title{Terminate an RSelenium WebDriver.}
\usage{
terminate_client(WebDriver = remDr, WebDriverName = "remDr", verbose = FALSE)
}
\arguments{
\item{WebDriver}{Active binding object containing the active WebDriver.}

\item{WebDriverName}{Name of the active binding object as characters.}

\item{verbose}{TRUE/FALSE to generate query window prior to running.}
}
\description{
Terminate the active WebDriver. Closes client window, ends server,
stops Java processes.
}
