% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terminate_client.R
\name{terminate_client}
\alias{terminate_client}
\title{Terminate an RSelenium WebDriver client}
\usage{
terminate_client(verbose = FALSE)
}
\arguments{
\item{verbose}{TRUE/FALSE generate query window prior to running.}
}
\description{
Terminate the default RSelenium client remDr. Closes window, ends server,
stops Java processes.
}
