% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terminate_client_nodriver.R
\name{terminate_client_nodriver}
\alias{terminate_client_nodriver}
\title{Terminate an RSelenium WebDriver client that was orphaned between sessions.}
\usage{
terminate_client_nodriver(verbose = FALSE, ping_check = FALSE, port = 4567L)
}
\arguments{
\item{verbose}{TRUE/FALSE generate query window prior to running.}

\item{ping_check}{TRUE/FALSE to use pingr to check the designated port after terminating Java.}

\item{port}{The port used for the WebDriver, to be checked if ping_check is TRUE.}
}
\description{
Terminate the default RSelenium client remDr if there is no longer an active
binding object. Stops Java processes. Use this if the active binding object
was removed. Client window must be closed manually.
}
